functions/reindex/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-secret-manager==2.*
PyGithub==2.*
langchain==0.3.*
langchain-huggingface==0.1.*
langchain-chroma==0.1.*
sentence-transformers==3.*
chromadb==0.5.*
google-genai